@@ -1,11 +1,11 @@
 name: klippa_scanner_sdk
 description: Allows you to do document scanning with the Klippa Scanner SDK from Flutter apps.
-version: 0.1.0
+version: 0.1.1
 homepage: https://github.com/klippa-app/flutter-klippa-scanner-sdk
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.20.0 < 2.0.0"
+  sdk: '>=2.12.0 <4.0.0'
+  flutter: '>=1.20.0'
 
 dependencies:
   flutter:
